CL-643: TraversalResult moved to public API. DAO API cleaned from duplicates.
