+++ common/src/main/java/com/gentics/mesh/core/db/AbstractEdgeFrame.java
@@ -1,8 +1,8 @@
 package com.gentics.mesh.core.db;
 
+import com.gentics.mesh.core.result.TraversalResult;
 import com.gentics.mesh.madl.frame.EdgeFrame;
 import com.gentics.mesh.madl.frame.VertexFrame;
-import com.gentics.mesh.madl.traversal.TraversalResult;
 import com.syncleus.ferma.AbstractElementFrame;
 import com.syncleus.ferma.FramedGraph;
 import com.syncleus.ferma.traversals.EdgeTraversal;
