+++ core/src/main/java/com/gentics/mesh/core/data/container/impl/MicroschemaContainerVersionImpl.java
@@ -32,7 +32,7 @@ import com.gentics.mesh.core.rest.microschema.impl.MicroschemaResponse;
 import com.gentics.mesh.core.rest.schema.MicroschemaReference;
 import com.gentics.mesh.core.rest.schema.impl.MicroschemaReferenceImpl;
 import com.gentics.mesh.core.result.Result;
-import com.gentics.mesh.madl.traversal.TraversalResult;
+import com.gentics.mesh.core.result.TraversalResult;
 import com.syncleus.ferma.ElementFrame;
 
 /**
