+++ madl/madl-ferma/src/main/java/com/gentics/madl/frame/AbstractVertexFrame.java
@@ -7,11 +7,11 @@ import com.gentics.madl.traversal.RawTraversalResult;
 import com.gentics.madl.tx.BaseTransaction;
 import com.gentics.madl.tx.Tx;
 import com.gentics.mesh.core.result.Result;
+import com.gentics.mesh.core.result.TraversalResult;
 import com.gentics.mesh.madl.frame.EdgeFrame;
 import com.gentics.mesh.madl.frame.ElementFrame;
 import com.gentics.mesh.madl.frame.VertexFrame;
 import com.gentics.mesh.madl.tp3.mock.GraphTraversal;
-import com.gentics.mesh.madl.traversal.TraversalResult;
 import com.syncleus.ferma.FramedGraph;
 import com.tinkerpop.blueprints.Direction;
 import com.tinkerpop.blueprints.Element;
