+++ mdm/api/src/main/java/com/gentics/mesh/core/data/dao/Dao.java
@@ -1,8 +1,14 @@
 package com.gentics.mesh.core.data.dao;
 
+import static com.gentics.mesh.core.data.perm.InternalPermission.READ_PERM;
 import static com.gentics.mesh.core.rest.error.Errors.error;
 import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
 import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
+import static org.apache.commons.lang3.StringUtils.isEmpty;
+
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Stream;
 
 import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBaseElement;
@@ -15,6 +21,7 @@ import com.gentics.mesh.core.rest.common.GenericRestResponse;
 import com.gentics.mesh.core.rest.common.PermissionInfo;
 import com.gentics.mesh.core.rest.common.RestModel;
 import com.gentics.mesh.core.result.Result;
+import com.gentics.mesh.core.result.TraversalResult;
 
 /**
  * General DAO.
@@ -22,25 +29,6 @@ import com.gentics.mesh.core.result.Result;
  * @param <T>
  */
 public interface Dao<T extends HibBaseElement> {
-
-	/**
-	 * Return the permission info for the given element and role.
-	 * 
-	 * @param element
-	 * @param ac
-	 * @param roleUuid
-	 * @return
-	 */
-	PermissionInfo getRolePermissions(HibCoreElement<? extends RestModel> element, InternalActionContext ac, String roleUuid);
-
-	/**
-	 * Return the roles which grant the given permission on the element.
-	 * 
-	 * @param element
-	 * @param perm
-	 * @return
-	 */
-	Result<? extends HibRole> getRolesWithPerm(T element, InternalPermission perm);
 
 	/**
 	 * Set the role permissionf for the given element.
@@ -90,5 +78,48 @@ public interface Dao<T extends HibBaseElement> {
 		} else {
 			throw error(FORBIDDEN, "error_missing_perm", elementUuid, perm.getRestPerm().getName());
 		}
+	}
+
+	/**
+	 * Return the roles which grant the given permission on the element.
+	 * 
+	 * @param element
+	 * @param perm
+	 * @return
+	 */
+	default Result<? extends HibRole> getRolesWithPerm(T element, InternalPermission perm) {
+		RoleDao roleDao = Tx.get().roleDao();
+		Set<String> roleUuids = roleDao.getRoleUuidsForPerm(element, perm);
+		Stream<String> stream = roleUuids == null
+			? Stream.empty()
+			: roleUuids.stream();
+		return new TraversalResult<>(stream
+			.map(roleDao::findByUuid)
+			.filter(Objects::nonNull));
+	}
+
+	/**
+	 * Return the permission info for the given element and role.
+	 * 
+	 * @param element
+	 * @param ac
+	 * @param roleUuid
+	 * @return
+	 */
+	default PermissionInfo getRolePermissions(HibCoreElement<? extends RestModel> element, InternalActionContext ac, String roleUuid) {
+		if (!isEmpty(roleUuid)) {
+			RoleDao roleDao = Tx.get().roleDao();
+			HibRole role = roleDao.loadObjectByUuid(ac, roleUuid, READ_PERM);
+			if (role != null) {
+				PermissionInfo permissionInfo = new PermissionInfo();
+				Set<InternalPermission> permSet = roleDao.getPermissions(role, element);
+				for (InternalPermission permission : permSet) {
+					permissionInfo.set(permission.getRestPerm(), true);
+				}
+				permissionInfo.setOthers(false);
+				return permissionInfo;
+			}
+		}
+		return null;
 	}
 }
